testPresignedGetUrlS3 working! Refactoring
