+++ ecomarkets/src/main/java/ecomarkets/domain/register/CPF.java
@@ -3,9 +3,9 @@ package ecomarkets.domain.register;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record CPF (Integer cpf){
+public record CPF (String cpf){
 
-    public static CPF of(Integer cpf){
+    public static CPF of(String cpf){
         return new CPF(cpf);
     }
