+++ ecomarkets/src/main/java/ecomarkets/domain/register/Telephone.java
@@ -3,9 +3,9 @@ package ecomarkets.domain.register;
 import jakarta.persistence.Embeddable;
 
 @Embeddable
-public record Telephone (Integer areaCode, Integer number){
+public record Telephone (String areaCode, String number){
 
-    public static Telephone of(Integer areaCode, Integer number){
+    public static Telephone of(String areaCode, String number){
         return new Telephone(areaCode, number);
     }
